Added rendering for game description text
* Moved GameSelectButton to its own class file
* Added rendering for desc_text with line wrapping
